Projects/Untitled Folder/mysite/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-heroku==0.3.1
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
gunicorn==20.0.4
jsmin==2.2.2
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
